fix(phase6): resolve compilation errors in services and components
- Fix decimal/double comparison in AnomalyDetectionService line 255
- Fix nullable int conversion in AnalyticsNotificationService line 105
- Fix MaxWidth.Lg → MaxWidth.Large in DashboardPreferences.razor
- Add TypeParameter="CustomReportTemplate" to MudDataGrid in DashboardPreferences
- Fix Severity.Info enum reference in DashboardPreferences.razor
- Add using statement for CreateReportTemplateDialog in DashboardPreferences.razor

+++ src/MyDesk.Web/Services/AnomalyDetectionService.cs
@@ -252,7 +252,7 @@ public class AnomalyDetectionService
 
         if (totalSpent > 0)
         {
-            foreach (var category in categoryTotals.Where(c => (c.Value / totalSpent) > 0.5))
+            foreach (var category in categoryTotals.Where(c => (c.Value / totalSpent) > (decimal)0.5))
             {
                 anomalies.Add(new ExpenseAnomaly
                 {
